Add empty line when no imports already defined
In cases where you have no imports defined yet and you append one it
would place the new import at line 1. This meant that is butted up right
next to the code. This commit will add an extra blank line in this case.

If anyone doesn't like this let me know and we can add an option to turn
it off.